+++ src/Repository/TaskRepository.php
@@ -12,4 +12,15 @@ class TaskRepository extends AbstractRepository implements TaskRepositoryContrac
     {
         parent::__construct($entityManager, Task::class);
     }
+
+    public function findAllExpired(): array
+    {
+        $queryBuilder = $this->entityManager->createQueryBuilder();
+        $queryBuilder
+            ->select('t')
+            ->from(Task::class, 't')
+            ->where('t.deadline < :deadline AND t.status != \'ended\'')
+            ->setParameter('deadline', date('Y-m-d'));
+        return $queryBuilder->getQuery()->getResult();
+    }
 }